requirements.txt
Flask==2.3.2
pymongo==4.6.1
bcrypt==4.1.2
PyJWT==2.8.0
flasgger==0.9.7.1
pytest==8.2.0
python-dotenv==1.0.1